+++ window_scene/session/host/include/scene_session.h
@@ -164,7 +164,7 @@ public:
     bool IsDecorEnable();
     void UpdateNativeVisibility(bool visible);
     void SetPrivacyMode(bool isPrivacy);
-	void SetSystemSceneBGAlpha(float alpha);
+    void SetSystemSceneBGAlpha(double alpha);
     bool IsVisible() const;
     bool IsFloatingWindowAppType() const;
     void DumpSessionElementInfo(const std::vector<std::string>& params);
